moved telegram logging to telegram controller
