+++ opendemic/fulfillment/command.py
@@ -5,7 +5,6 @@ from opendemic.channels.telegram import TelegramCommand, get_telegram_bot_instan
 	make_reply_keyboard_markup, get_telegram_menu
 from telebot import types
 from opendemic.models.human import Human
-from opendemic.logging.action import log_sent_message, log_action
 from helpers.url import url_add_params, compose_client_url
 import random
