+++ opendemic/models/human.py
@@ -9,12 +9,8 @@ from helpers.id import verify_uuid_regex
 from opendemic.database.sql_db import RDBManager
 from opendemic.channels.telegram import get_telegram_bot_instance, get_telegram_menu, make_reply_keyboard_markup
 from helpers.formatting import quote_wrap, mysql_db_format_value
-from helpers.datetime import datetime_to_mysql
-from opendemic.logging.action import log_action
 import datetime
-import json
 import uuid
-import pandas as pd
 import os
